Client: added UserData type
